refactor(LabelEditor): make label editor generic
